refactor(backend): align modules with domain architecture - split application, matching, and HTTP server concerns out of apartment - move application queries and use cases into dedicated module - add group and closure module placeholders for future domain work
