+++ backend/internal/apartment/httpadapter/handler.go
@@ -11,14 +11,14 @@ import (
 	"github.com/isw2-unileon/proyect-scaffolding/backend/internal/apartment"
 	apartmentservice "github.com/isw2-unileon/proyect-scaffolding/backend/internal/apartment/service"
 	authservice "github.com/isw2-unileon/proyect-scaffolding/backend/internal/auth/service"
-	"github.com/isw2-unileon/proyect-scaffolding/backend/internal/httpauth"
 	profileservice "github.com/isw2-unileon/proyect-scaffolding/backend/internal/profile/service"
 )
 
 type handler struct {
-	authService      *authservice.Service
-	profileService   *profileservice.Service
-	apartmentService *apartmentservice.Service
+	authService        *authservice.Service
+	profileService     *profileservice.Service
+	apartmentService   *apartmentservice.Service
+	extractBearerToken func(string) (string, error)
 }
 
 type createApartmentRequest struct {
@@ -80,47 +80,11 @@ type apartmentRulesResponse struct {
 	PreferredSchedule      string `json:"preferred_schedule"`
 }
 
-type applyApartmentResponse struct {
-	ApplicationID string `json:"application_id"`
-	Status        string `json:"status"`
-}
-
-type interestedTenantResponse struct {
-	UserID        string `json:"user_id"`
-	Name          string `json:"name"`
-	Age           int    `json:"age"`
-	Studies       string `json:"studies"`
-	AvatarURL     string `json:"avatar_url"`
-	Compatibility int    `json:"compatibility"`
-}
-
-type tenantApplicationResponse struct {
-	ID            string `json:"id"`
-	ApartmentID   string `json:"apartment_id"`
-	PropertyTitle string `json:"property_title"`
-	OwnerName     string `json:"owner_name"`
-	Address       string `json:"address"`
-	ImageURL      string `json:"image_url"`
-	Places        int    `json:"places"`
-	Size          int    `json:"size"`
-	Bathrooms     int    `json:"bathrooms"`
-	Status        string `json:"status"`
-	CreatedAt     string `json:"created_at"`
-	DateLabel     string `json:"date_label"`
-	Compatibility int    `json:"compatibility"`
-	RequestType   string `json:"request_type"`
-	StatusMessage string `json:"status_message"`
-}
-
 // RegisterRoutes wires apartment endpoints into the API router.
-func RegisterRoutes(api *gin.RouterGroup, authService *authservice.Service, profileService *profileservice.Service, apartmentService *apartmentservice.Service) {
-	h := &handler{authService: authService, profileService: profileService, apartmentService: apartmentService}
+func RegisterRoutes(api *gin.RouterGroup, authService *authservice.Service, profileService *profileservice.Service, apartmentService *apartmentservice.Service, extractBearerToken func(string) (string, error)) {
+	h := &handler{authService: authService, profileService: profileService, apartmentService: apartmentService, extractBearerToken: extractBearerToken}
 	api.GET("/apartments", h.listAvailableApartments)
 	api.GET("/apartments/:id", h.getApartmentDetail)
-	api.POST("/apartments/:id/applications", h.applyToApartment)
-	api.POST("/applications/:id/cancel", h.cancelTenantApplication)
-	api.GET("/apartments/:id/interested", h.listInterestedTenants)
-	api.GET("/tenant/applications", h.listTenantApplications)
 	api.GET("/owner/apartments", h.listOwnerApartments)
 	api.POST("/apartments", h.createApartment)
 }
@@ -242,132 +206,6 @@ func (h *handler) getApartmentDetail(c *gin.Context) {
 	})
 }
 
-func (h *handler) applyToApartment(c *gin.Context) {
-	tenantID, role, ok := h.resolveUserAndRole(c)
-	if !ok {
-		return
-	}
-	apartmentID := strings.TrimSpace(c.Param("id"))
-	if apartmentID == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "apartment id is required"})
-		return
-	}
-	applicationID, err := h.apartmentService.ApplyToApartment(c.Request.Context(), apartmentID, tenantID, role)
-	if err != nil {
-		if errors.Is(err, apartmentservice.ErrTenantRequired) {
-			c.JSON(http.StatusForbidden, gin.H{"error": "apartment applications are only available for tenant users"})
-			return
-		}
-		if errors.Is(err, apartmentservice.ErrApartmentNotFound) {
-			c.JSON(http.StatusNotFound, gin.H{"error": "apartment not found"})
-			return
-		}
-		if errors.Is(err, apartmentservice.ErrApartmentFull) {
-			c.JSON(http.StatusConflict, gin.H{"error": "apartment is full"})
-			return
-		}
-		if errors.Is(err, apartmentservice.ErrApplicationAlreadyExists) {
-			c.JSON(http.StatusConflict, gin.H{"error": "application already exists"})
-			return
-		}
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "could not create application"})
-		return
-	}
-	c.JSON(http.StatusCreated, applyApartmentResponse{ApplicationID: applicationID, Status: "pending"})
-}
-
-func (h *handler) listInterestedTenants(c *gin.Context) {
-	apartmentID := strings.TrimSpace(c.Param("id"))
-	if apartmentID == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "apartment id is required"})
-		return
-	}
-	interested, err := h.apartmentService.ListInterestedTenants(c.Request.Context(), apartmentID)
-	if err != nil {
-		if errors.Is(err, apartmentservice.ErrApartmentNotFound) {
-			c.JSON(http.StatusNotFound, gin.H{"error": "apartment not found"})
-			return
-		}
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "could not load interested tenants"})
-		return
-	}
-	response := make([]interestedTenantResponse, 0, len(interested))
-	for _, item := range interested {
-		response = append(response, interestedTenantResponse{
-			UserID:        item.UserID,
-			Name:          item.Name,
-			Age:           item.Age,
-			Studies:       item.Studies,
-			AvatarURL:     item.AvatarURL,
-			Compatibility: item.Compatibility,
-		})
-	}
-	c.JSON(http.StatusOK, gin.H{"tenants": response})
-}
-
-func (h *handler) cancelTenantApplication(c *gin.Context) {
-	tenantID, role, ok := h.resolveUserAndRole(c)
-	if !ok {
-		return
-	}
-	applicationID := strings.TrimSpace(c.Param("id"))
-	if applicationID == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "application id is required"})
-		return
-	}
-	err := h.apartmentService.CancelTenantApplication(c.Request.Context(), applicationID, tenantID, role)
-	if err != nil {
-		if errors.Is(err, apartmentservice.ErrTenantRequired) {
-			c.JSON(http.StatusForbidden, gin.H{"error": "applications are only available for tenant users"})
-			return
-		}
-		if errors.Is(err, apartmentservice.ErrApplicationNotCancelable) {
-			c.JSON(http.StatusConflict, gin.H{"error": "application is not cancelable"})
-			return
-		}
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "could not cancel application"})
-		return
-	}
-	c.JSON(http.StatusOK, gin.H{"message": "application cancelled"})
-}
-
-func (h *handler) listTenantApplications(c *gin.Context) {
-	tenantID, role, ok := h.resolveUserAndRole(c)
-	if !ok {
-		return
-	}
-	applications, err := h.apartmentService.ListTenantApplications(c.Request.Context(), tenantID, role)
-	if err != nil {
-		if errors.Is(err, apartmentservice.ErrTenantRequired) {
-			c.JSON(http.StatusForbidden, gin.H{"error": "applications are only available for tenant users"})
-			return
-		}
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "could not load tenant applications"})
-		return
-	}
-	response := make([]tenantApplicationResponse, 0, len(applications))
-	for _, item := range applications {
-		response = append(response, tenantApplicationResponse{
-			ID:            item.ID,
-			ApartmentID:   item.ApartmentID,
-			PropertyTitle: item.PropertyTitle,
-			OwnerName:     item.OwnerName,
-			Address:       item.Address,
-			ImageURL:      item.ImageURL,
-			Places:        item.Places,
-			Size:          item.Size,
-			Bathrooms:     item.Bathrooms,
-			Status:        item.Status,
-			CreatedAt:     item.CreatedAt,
-			DateLabel:     item.DateLabel,
-			Compatibility: item.CompatibilityScore,
-			RequestType:   item.RequestType,
-			StatusMessage: item.StatusMessage,
-		})
-	}
-	c.JSON(http.StatusOK, gin.H{"applications": response})
-}
-
 func (h *handler) listOwnerApartments(c *gin.Context) {
 	ownerID, role, ok := h.resolveUserAndRole(c)
 	if !ok {
@@ -387,7 +225,7 @@ func (h *handler) listOwnerApartments(c *gin.Context) {
 }
 
 func (h *handler) resolveUserAndRole(c *gin.Context) (string, string, bool) {
-	accessToken, err := httpauth.ExtractBearerToken(c.GetHeader("Authorization"))
+	accessToken, err := h.extractBearerToken(c.GetHeader("Authorization"))
 	if err != nil {
 		c.JSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
 		return "", "", false
