+++ backend/internal/apartment/service/service_test.go
@@ -6,26 +6,21 @@ import (
 	"testing"
 
 	"github.com/isw2-unileon/proyect-scaffolding/backend/internal/apartment"
+	"github.com/isw2-unileon/proyect-scaffolding/backend/internal/profile"
 )
 
 type fakeApartmentRepository struct {
-	createdDescription            string
-	createdStatus                 string
-	createdApplicationApartmentID string
-	createdApplicationTenantID    string
-	createdCompatibilityScore     int
-	hasActiveApplication          bool
+	createdDescription string
+	createdStatus      string
+
 	applicationForApartmentID     string
 	applicationForApartmentStatus string
-	cancelApplicationUpdated      bool
 
-	ownerApartments    []apartment.Apartment
-	tenantApartments   []apartment.Apartment
-	apartmentByID      *apartment.Apartment
-	apartmentRules     *apartment.Rules
-	tenantProfile      *apartment.TenantProfile
-	interestedTenants  []apartment.InterestedTenantCandidate
-	tenantApplications []apartment.TenantApplication
+	ownerApartments  []apartment.Apartment
+	tenantApartments []apartment.Apartment
+	apartmentByID    *apartment.Apartment
+	apartmentRules   *apartment.Rules
+	tenantProfile    *profile.TenantProfile
 }
 
 func (f *fakeApartmentRepository) CreateApartment(ctx context.Context, ownerID string, input apartment.CreateApartmentInput) (string, int, error) {
@@ -69,11 +64,11 @@ func (f *fakeApartmentRepository) GetApartmentRules(ctx context.Context, apartme
 	}, nil
 }
 
-func (f *fakeApartmentRepository) GetTenantProfileByUserID(ctx context.Context, userID string) (*apartment.TenantProfile, error) {
+func (f *fakeApartmentRepository) GetTenantProfileByUserID(ctx context.Context, userID string) (*profile.TenantProfile, error) {
 	if f.tenantProfile != nil {
 		return f.tenantProfile, nil
 	}
-	return &apartment.TenantProfile{
+	return &profile.TenantProfile{
 		UserID:        userID,
 		BudgetMin:     300,
 		BudgetMax:     450,
@@ -86,53 +81,8 @@ func (f *fakeApartmentRepository) GetTenantProfileByUserID(ctx context.Context, 
 	}, nil
 }
 
-func (f *fakeApartmentRepository) HasActiveApplication(ctx context.Context, apartmentID, tenantID string) (bool, error) {
-	return f.hasActiveApplication, nil
-}
-
-func (f *fakeApartmentRepository) CreateTenantApplication(ctx context.Context, apartmentID, tenantID string, compatibilityScore int) (string, error) {
-	f.createdApplicationApartmentID = apartmentID
-	f.createdApplicationTenantID = tenantID
-	f.createdCompatibilityScore = compatibilityScore
-	return "application-1", nil
-}
-
 func (f *fakeApartmentRepository) GetTenantApplicationForApartment(ctx context.Context, apartmentID, tenantID string) (string, string, error) {
 	return f.applicationForApartmentID, f.applicationForApartmentStatus, nil
-}
-
-func (f *fakeApartmentRepository) CancelTenantApplication(ctx context.Context, applicationID, tenantID string) (bool, error) {
-	return f.cancelApplicationUpdated, nil
-}
-
-func (f *fakeApartmentRepository) ListInterestedTenants(ctx context.Context, apartmentID string) ([]apartment.InterestedTenantCandidate, error) {
-	if f.interestedTenants != nil {
-		return f.interestedTenants, nil
-	}
-	return []apartment.InterestedTenantCandidate{{
-		InterestedTenant: apartment.InterestedTenant{UserID: "tenant-2", Name: "Laura", Age: 21, Studies: "Veterinaria"},
-		BudgetMin:        320,
-		BudgetMax:        460,
-		PreferredArea:    "centro",
-		Pets:             false,
-		Smoking:          false,
-		NoiseLevel:       "moderate",
-		Cleanliness:      "normal",
-		WorkSchedule:     "flexible",
-	}}, nil
-}
-
-func (f *fakeApartmentRepository) ListTenantApplications(ctx context.Context, tenantID string) ([]apartment.TenantApplication, error) {
-	if f.tenantApplications != nil {
-		return f.tenantApplications, nil
-	}
-	return []apartment.TenantApplication{{
-		ID:            "application-1",
-		ApartmentID:   "apartment-1",
-		PropertyTitle: "Flat",
-		Status:        "PENDING_OWNER",
-		CreatedAt:     "2026-05-23",
-	}}, nil
 }
 
 type fakeImageSigner struct {
@@ -150,7 +100,7 @@ func (f *fakeImageSigner) CreateSignedURL(ctx context.Context, bucket string, pa
 
 func TestCreateApartmentAcceptsRepositoryInterfaceAndPreparesPersistenceData(t *testing.T) {
 	repo := &fakeApartmentRepository{}
-	svc := NewService(repo, nil)
+	svc := NewService(repo, nil, repo, repo)
 
 	result, err := svc.CreateApartment(context.Background(), "owner-1", "owner", apartment.CreateApartmentInput{
 		Title:         " Flat ",
@@ -179,7 +129,8 @@ func TestCreateApartmentAcceptsRepositoryInterfaceAndPreparesPersistenceData(t *
 }
 
 func TestCreateApartmentRejectsNonOwnerRole(t *testing.T) {
-	svc := NewService(&fakeApartmentRepository{}, nil)
+	repo := &fakeApartmentRepository{}
+	svc := NewService(repo, nil, repo, repo)
 
 	_, err := svc.CreateApartment(context.Background(), "tenant-1", "tenant", apartment.CreateApartmentInput{})
 	if !errors.Is(err, ErrOwnerRequired) {
@@ -188,7 +139,8 @@ func TestCreateApartmentRejectsNonOwnerRole(t *testing.T) {
 }
 
 func TestListAvailableApartmentsReturnsTenantVisibleListings(t *testing.T) {
-	svc := NewService(&fakeApartmentRepository{}, nil)
+	repo := &fakeApartmentRepository{}
+	svc := NewService(repo, nil, repo, repo)
 
 	apartments, err := svc.ListAvailableApartments(context.Background())
 	if err != nil {
@@ -211,7 +163,7 @@ func TestListOwnerApartmentsSignsImagePaths(t *testing.T) {
 		}},
 	}
 	signer := &fakeImageSigner{}
-	svc := NewService(repo, signer)
+	svc := NewService(repo, signer, repo, repo)
 
 	apartments, err := svc.ListOwnerApartments(context.Background(), "owner-1", "owner")
 	if err != nil {
@@ -243,7 +195,7 @@ func TestListAvailableApartmentsSignsTenantImagePaths(t *testing.T) {
 		}},
 	}
 	signer := &fakeImageSigner{}
-	svc := NewService(repo, signer)
+	svc := NewService(repo, signer, repo, repo)
 
 	apartments, err := svc.ListAvailableApartments(context.Background())
 	if err != nil {
@@ -257,7 +209,7 @@ func TestListAvailableApartmentsSignsTenantImagePaths(t *testing.T) {
 
 func TestGetApartmentDetailForTenantReturnsCompatibility(t *testing.T) {
 	repo := &fakeApartmentRepository{applicationForApartmentID: "application-1", applicationForApartmentStatus: "PENDING_OWNER"}
-	svc := NewService(repo, nil)
+	svc := NewService(repo, nil, repo, repo)
 
 	detail, err := svc.GetApartmentDetailForTenant(context.Background(), "apartment-1", "tenant-1", "tenant")
 	if err != nil {
@@ -271,83 +223,5 @@ func TestGetApartmentDetailForTenantReturnsCompatibility(t *testing.T) {
 	}
 	if detail.CanApply {
 		t.Fatalf("CanApply = %t, want false", detail.CanApply)
-	}
-}
-
-func TestApplyToApartmentCreatesTenantApplication(t *testing.T) {
-	repo := &fakeApartmentRepository{}
-	svc := NewService(repo, nil)
-
-	applicationID, err := svc.ApplyToApartment(context.Background(), "apartment-1", "tenant-1", "tenant")
-	if err != nil {
-		t.Fatalf("ApplyToApartment returned error: %v", err)
-	}
-	if applicationID != "application-1" {
-		t.Fatalf("applicationID = %q, want application-1", applicationID)
-	}
-	if repo.createdCompatibilityScore <= 0 {
-		t.Fatalf("createdCompatibilityScore = %d, want > 0", repo.createdCompatibilityScore)
-	}
-}
-
-func TestApplyToApartmentRejectsDuplicateActiveApplication(t *testing.T) {
-	repo := &fakeApartmentRepository{hasActiveApplication: true}
-	svc := NewService(repo, nil)
-
-	_, err := svc.ApplyToApartment(context.Background(), "apartment-1", "tenant-1", "tenant")
-	if !errors.Is(err, ErrApplicationAlreadyExists) {
-		t.Fatalf("err = %v, want %v", err, ErrApplicationAlreadyExists)
-	}
-}
-
-func TestListInterestedTenantsCalculatesCompatibility(t *testing.T) {
-	repo := &fakeApartmentRepository{}
-	svc := NewService(repo, nil)
-
-	result, err := svc.ListInterestedTenants(context.Background(), "apartment-1")
-	if err != nil {
-		t.Fatalf("ListInterestedTenants returned error: %v", err)
-	}
-	if len(result) != 1 {
-		t.Fatalf("len(result) = %d, want 1", len(result))
-	}
-	if result[0].Compatibility <= 0 {
-		t.Fatalf("Compatibility = %d, want > 0", result[0].Compatibility)
-	}
-}
-
-func TestListTenantApplicationsMapsStatus(t *testing.T) {
-	repo := &fakeApartmentRepository{}
-	svc := NewService(repo, nil)
-
-	result, err := svc.ListTenantApplications(context.Background(), "tenant-1", "tenant")
-	if err != nil {
-		t.Fatalf("ListTenantApplications returned error: %v", err)
-	}
-	if len(result) != 1 {
-		t.Fatalf("len(result) = %d, want 1", len(result))
-	}
-	if result[0].Status != "pending" {
-		t.Fatalf("Status = %q, want pending", result[0].Status)
-	}
-}
-
-func TestCancelTenantApplicationRejectsNotCancelable(t *testing.T) {
-	repo := &fakeApartmentRepository{cancelApplicationUpdated: false}
-	svc := NewService(repo, nil)
-
-	err := svc.CancelTenantApplication(context.Background(), "application-1", "tenant-1", "tenant")
-	if !errors.Is(err, ErrApplicationNotCancelable) {
-		t.Fatalf("err = %v, want %v", err, ErrApplicationNotCancelable)
-	}
-}
-
-func TestCancelTenantApplicationUpdatesPendingApplication(t *testing.T) {
-	repo := &fakeApartmentRepository{cancelApplicationUpdated: true}
-	svc := NewService(repo, nil)
-
-	err := svc.CancelTenantApplication(context.Background(), "application-1", "tenant-1", "tenant")
-	if err != nil {
-		t.Fatalf("CancelTenantApplication returned error: %v", err)
 	}
 }
