+++ backend/internal/closure/closure.go
@@ -1,0 +1,1 @@
+package closure
