+++ backend/internal/group/group.go
@@ -1,0 +1,1 @@
+package group
