+++ backend/internal/httpserver/router.go
@@ -1,4 +1,4 @@
-package httpapi
+package httpserver
 
 import (
 	"net/http"
@@ -6,6 +6,8 @@ import (
 	"github.com/gin-gonic/gin"
 	apartmenthttp "github.com/isw2-unileon/proyect-scaffolding/backend/internal/apartment/httpadapter"
 	apartmentservice "github.com/isw2-unileon/proyect-scaffolding/backend/internal/apartment/service"
+	applicationhttp "github.com/isw2-unileon/proyect-scaffolding/backend/internal/application/httpadapter"
+	applicationservice "github.com/isw2-unileon/proyect-scaffolding/backend/internal/application/service"
 	authhttp "github.com/isw2-unileon/proyect-scaffolding/backend/internal/auth/httpadapter"
 	authservice "github.com/isw2-unileon/proyect-scaffolding/backend/internal/auth/service"
 	"github.com/isw2-unileon/proyect-scaffolding/backend/internal/platform/config"
@@ -14,7 +16,7 @@ import (
 )
 
 // NewRouter builds the HTTP API router.
-func NewRouter(cfg *config.Config, authService *authservice.Service, profileService *profileservice.Service, apartmentService *apartmentservice.Service) *gin.Engine {
+func NewRouter(cfg *config.Config, authService *authservice.Service, profileService *profileservice.Service, apartmentService *apartmentservice.Service, applicationService *applicationservice.Service) *gin.Engine {
 	gin.SetMode(cfg.GinMode)
 	r := gin.New()
 	r.Use(gin.Logger(), gin.Recovery(), corsMiddleware(cfg.CORSAllowOrigin))
@@ -31,10 +33,13 @@ func NewRouter(cfg *config.Config, authService *authservice.Service, profileServ
 	if authService == nil || profileService == nil {
 		return r
 	}
-	authhttp.RegisterRoutes(api, authService, cfg.FrontendURL)
-	profilehttp.RegisterRoutes(api, authService, profileService)
+	authhttp.RegisterRoutes(api, authService, cfg.FrontendURL, ExtractBearerToken)
+	profilehttp.RegisterRoutes(api, authService, profileService, ExtractBearerToken)
 	if apartmentService != nil {
-		apartmenthttp.RegisterRoutes(api, authService, profileService, apartmentService)
+		apartmenthttp.RegisterRoutes(api, authService, profileService, apartmentService, ExtractBearerToken)
+	}
+	if applicationService != nil {
+		applicationhttp.RegisterRoutes(api, authService, profileService, applicationService, ExtractBearerToken)
 	}
 	return r
 }
